+++ changelog.md
@@ -5,6 +5,10 @@ Version History
 -----------
 Here you'll find information on key changes, bug fixes and improvements made by version.
 
+### What's new in version 1907.1?
+* MultiNode fixes
+* Default Provider Subscription fixes
+
 ### What's new in version 1907?
 * Updated to support ASDK 1.1907.0.20
 * Initial support for MultiNode deployments for POC purposes only